refactor: :wastebasket: Product Attributes are a part of Custom Attributes
