+++ src/MagentoStorageServiceProvider.php
@@ -83,7 +83,6 @@ class MagentoStorageServiceProvider extends ServiceProvider
             Console\SyncMagentoProductCategoriesCommand::class,
             Console\SyncMagentoCustomAttributeTypesCommand::class,
             Console\SyncMagentoWebsitesCommand::class,
-            Console\SyncMagentoProductAttributesCommand::class,
         ]);
     }
 }
