--- src/Models/MagentoProductAttribute.php
@@ -1,43 +1,0 @@
-<?php
-
-namespace Grayloon\MagentoStorage\Models;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-
-class MagentoProductAttribute extends Model
-{
-    use HasFactory;
-
-    /**
-     * The attributes that are mass assignable.
-     *
-     * @var array
-     */
-    protected $fillable = [
-        'id',
-        'name',
-        'code',
-        'position',
-        'default_value',
-        'type',
-        'synced_at',
-    ];
-
-    /**
-     * The attributes that should be mutated to dates.
-     *
-     * @var array
-     */
-    protected $dates = ['created_at', 'updated_at', 'synced_at'];
-
-    /**
-     * The attribute can have many Option attributes.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\HasMany
-     */
-    public function options()
-    {
-        return $this->hasMany(MagentoProductAttributeOption::class);
-    }
-}
