--- src/Models/MagentoProductAttributeOption.php
@@ -1,40 +1,0 @@
-<?php
-
-namespace Grayloon\MagentoStorage\Models;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-
-class MagentoProductAttributeOption extends Model
-{
-    use HasFactory;
-
-    /**
-     * The attributes that are mass assignable.
-     *
-     * @var array
-     */
-    protected $fillable = [
-        'label',
-        'value',
-        'magento_product_attribute_id',
-        'synced_at',
-    ];
-
-    /**
-     * The attributes that should be mutated to dates.
-     *
-     * @var array
-     */
-    protected $dates = ['created_at', 'updated_at', 'synced_at'];
-
-    /**
-     * The attribute belongs to Magento Product Attributes.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
-     */
-    public function attribute()
-    {
-        return $this->belongsTo(MagentoProductAttribute::class, 'magento_product_attribute_id');
-    }
-}
